feat: zod - organize spotify models and schemas
