+++ client/.env.example
@@ -1,0 +1,1 @@
+REACT_APP_SERVER_HOST=http://localhost:8888
