+++ client/src/types/spotify/enum/spotify.payload.enum.ts
@@ -1,0 +1,5 @@
+export enum SpotifyTimeRange {
+  LONG_TERM = 'long_term',
+  MEDIUM_TERM = 'medium_term',
+  SHORT_TERM = 'short_term',
+}
