+++ client/src/types/spotify/model/spotify.album.model.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { SpotifyAlbumSchema } from '../schema/spotify.album.schema';
+
+export type SpotifyAlbum = z.infer<typeof SpotifyAlbumSchema>;
