+++ client/src/types/spotify/model/spotify.artist.model.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { SpotifyArtistSchema } from '../schema/spotify.artist.schema';
+
+export type SpotifyArtist = z.infer<typeof SpotifyArtistSchema>;
