+++ client/src/types/spotify/model/spotify.payload.model.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { SpotifyPayloadSchema } from '../schema/spotify.payload.schema';
+
+export type SpotifyPayload = z.infer<typeof SpotifyPayloadSchema>;
