+++ client/src/types/spotify/model/spotify.profile.model.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { SpotifyProfileSchema } from '../schema/spotify.profile.schema';
+
+export type SpotifyProfile = z.infer<typeof SpotifyProfileSchema>;
