+++ client/src/types/spotify/model/spotify.user.model.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import {
+  SpotifyUserTopArtistsSchema,
+  SpotifyUserTopTracksSchema,
+} from '../schema/spotify.user.schema';
+
+export type SpotifyUserTopTracks = z.infer<typeof SpotifyUserTopTracksSchema>;
+export type SpotifyUserTopArtists = z.infer<typeof SpotifyUserTopArtistsSchema>;
